Added RegistrationInformationAnalyzer and code cleanup - seperated out the module, registration and startable diagnostics
